+++ openwave/xperiments/C_vector_field/__init__.py
@@ -1,5 +1,5 @@
 """
-D/ VECTOR-FIELD METHOD (analytical tools)
+C/ VECTOR-FIELD METHOD (analytical tools)
 
 Experiments exploring wave dynamics using vector-field method simulation.
 """
