+++ openwave/xperiments/C_vector_field/plots/charge_profile.png
