+++ openwave/xperiments/C_vector_field/research/01_wolff_lafreniere.md
