+++ openwave/xperiments/C_vector_field/research/plots/wave_amplitude_vs_radius.png
